# Conflicts:
#	DynaMind-GDALModules/scripts/GDALModules/__init__.py
#	DynaMind-GDALModules/scripts/GDALModules/dm_sftp_hotstart.py
#	DynaMind-GDALModules/src/drainage/swmmwriteandread.cpp
#	DynaMind-Performance-Assessment/scripts/DMPerformance/dm_sei.py

+++ DynaMind-GDALModules/scripts/GDALModules/dm_sftp_hotstart.py
@@ -3,7 +3,7 @@ from pydynamind import *
 
 import paramiko
 import time
-from osgeo import ogr
+from osgeo import gdal
 import uuid
 import os
 
@@ -74,9 +74,7 @@ class DM_Hoststart_SFTP(Module):
 
             self.close()
 
-
-            ds = ogr.Open(self.real_file_name)
-
+            ds = gdal.OpenEx(self.real_file_name, gdal.OF_VECTOR)
 
             if ds is None:
                 print "Open failed.\n"
@@ -192,6 +190,10 @@ class DM_Hoststart_SFTP(Module):
 
         def run(self):
             db = self.getGDALData("city")
+            if self.real_file_name == "":
+                log("File not downloaded", Error)
+                self.setStatus(MOD_EXECUTION_ERROR)
+                return
             db.setGDALDatabase(self.real_file_name)
             os.remove(self.real_file_name)
